libminiooni: remove --no-bouncer and --no-geoip
The removal of --no-geoip stems from the fact that now we require
measurements to be geolocated (see https://github.com/ooni/probe-engine/issues/974).

The removal of --no-bouncer is because that option was added some
time ago for MK compatibility, but we've later decided that we don't
want to provide any compatibility guarantee, so we can happily get
rid of this option, which is only adding extra complexity.