fix(saver, submitter): make sure we only log when needed (#1106)
While working on https://github.com/ooni/probe-engine/issues/1057, I noticed
that since recent changes in probe-engine that introduced Saver and Submitter,
we've got logging wrong. We should actually be logging only when we're are
submitting and/or saving, and we shan't be logging otherwise.

While there, make sure we continue to print the reportID, which is a very
useful piece of information to see emitted by miniooni.