feat(experiment): split InputRequired policy
We used to have InputRequired category, which meant that you were
providing input and, if there was not input, we would then use the
test-lists API to fetch input URLs.

This policy was used, however, improperly in some cases. As noted
in https://github.com/ooni/probe-engine/issues/1123, dnscheck would
for example use the test lists if no input was provided. Sadly,
though, such input does not make any sense when provided to dnscheck,
which leads to miniooni running very useless tests.

So, here's an improvement:

1. we rename InputRequired to InputOrQueryTestLists, which BTW
captures in an even better way the original intent

2. we introduce InputStrictlyRequired, where, if there is no
input, we just fail, and we apply it to dnscheck, run, and all
the cases in which it makes sense to use this new policy

Work done as part of https://github.com/ooni/probe-engine/issues/1115.