cleanup(report): remove report.close operation
The close operation is now a relic of the past that we support
only for the sake of legacy clients.

So, we can remove it and simplify the code a little bit.

While there, make sure we use an interface for the opened report
rather than a struct, so we can improve testing in the future.

Also, the OOID is currently not returned by any collector, hence
it does not make any sense to have code handling it.

Work done as cleanup after https://github.com/ooni/probe-engine/issues/948.