quicdialer: annotate doubts and refactor
Avoid names like quicdialer.QUICFoo. We're already into the quicdialer
package, so we can save typing the QUIC prefix.

Wrap lines that seem to long (a subjective call, of course).

It seems there were two similar mockable dialers, so unify them such
that we only have one. We gain a little in simplicity.