fix: allow mixing different nettests inside run
Since https://github.com/ooni/probe-engine/pull/1182 we can mix
urlgetter and dnscheck when using run. This is not awesome because
it violates implicit constraints in pipeline, explorer.

Specifically, a single reportID should always contain just a
single kind of nettests.

On top of that, submitting measurements using a reportID containing
the `run` experiment name was also very lame.

With this diff, we are making a decisive step forward by basically
changing the reportID we're using whenever we need it.

To this end, we change the way in which we submit. We're not using
the algorithm used by mobile resubmission.

This feels like part of https://github.com/ooni/probe-engine/issues/921.